Fix qwen_image_edit_plus_gguf loader: default load_inputs dtype to bfloat16 to match GGUF model

+++ qwen_image_edit_plus_gguf/pytorch/loader.py
@@ -117,7 +117,7 @@ class ModelLoader(ForgeModel):
 
     def load_inputs(self, **kwargs) -> Any:
         """Prepare sample inputs for the diffusion transformer."""
-        dtype = kwargs.get("dtype_override", torch.float32)
+        dtype = kwargs.get("dtype_override", torch.bfloat16)
         batch_size = kwargs.get("batch_size", 1)
 
         # From model config: in_channels=64 (img_in linear input dimension)
